% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nationalmusym-lookup.R
\name{create_nationalmusym_lookup}
\alias{create_nationalmusym_lookup}
\title{Create National MUSYM Lookup Snapshot}
\usage{
create_nationalmusym_lookup()
}
\value{
Data frame. With mukey, nationalmusym, and areasymbol columns.
}
\description{
Creates a snapshot of MUKEY:NATIONALMUSYM:AREASYMBOL from Soil Data Access for the current fiscal year, sorted by areasymbol and mukey.
}
