% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGOsnapshots-utils.R
\name{download_snapshot_asset}
\alias{download_snapshot_asset}
\title{Download Snapshot Asset}
\usage{
download_snapshot_asset(
  snapshot_id,
  fy,
  repo = "brownag/SSURGOsnapshots",
  force = FALSE
)
}
\arguments{
\item{snapshot_id}{Character. The ID of the snapshot.}

\item{fy}{Numeric. Fiscal year.}

\item{repo}{Character. GitHub repository. Default: \code{"brownag/SSURGOsnapshots"}.}

\item{force}{Logical. Force redownload even if cached. Default: \code{FALSE}.}
}
\value{
Data frame. From the downloaded CSV.
}
\description{
Download Snapshot Asset
}
