% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGOsnapshots-utils.R
\name{fiscal_year_from_date}
\alias{fiscal_year_from_date}
\title{Fiscal Year from Date}
\usage{
fiscal_year_from_date(current_date = Sys.Date(), fy_start_date = "2000-10-01")
}
\arguments{
\item{current_date}{Date. The date to calculate fiscal year for. Default: \code{Sys.Date()}.}

\item{fy_start_date}{Character. The start date of fiscal year in MM-DD format. Default: \code{"2000-10-01"}.}
}
\value{
Numeric. The fiscal year.
}
\description{
Fiscal Year from Date
}
