% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGOsnapshots-schema.R
\name{generate_draft_schema}
\alias{generate_draft_schema}
\title{Generate Draft Schema}
\usage{
generate_draft_schema(data, snapshot_id, fy)
}
\arguments{
\item{data}{Data frame. The data frame to generate schema from.}

\item{snapshot_id}{Character. The ID of the snapshot.}

\item{fy}{Numeric. Fiscal year.}
}
\value{
List. Invisible schema list written to file.
}
\description{
Generate Draft Schema
}
