% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGOsnapshots-package.R
\docType{package}
\name{SSURGOsnapshots-package}
\alias{SSURGOsnapshots}
\alias{SSURGOsnapshots-package}
\title{SSURGOsnapshots: SSURGO Data Snapshots}
\description{
Generates snapshots of Soil Survey Geographic Database (SSURGO) and related soil data from various sources, including Soil Data Access (SDA). Initial proof of concept is a 'nationalmusym' to 'MUKEY' lookup table.
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov} (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})

}
\keyword{internal}
